fix phone, connect argon2, validation, added api-response, custom logger

+++ src/main/java/burunzhuy/resource/user/UserResource.java
@@ -14,7 +14,7 @@ public class UserResource {
     private final String name;
     private final String secondName;
     private final String lastName;
-    private final Integer phone;
+    private final Long phone;
     private final Set<Role> roles;
 
     public UserResource(User user) {
